Removing prefix_search from testing
